Use defaultExecutor() to complete parent future [HZG-63] (#3472)
The following PRs

https://github.com/hazelcast/hazelcast/pull/23532
https://github.com/hazelcast/hazelcast/pull/23539

changed the executor that completes the parent future in
`DelegatingCompletableFuture` and `ClientDelegatingFuture` from
`defaultExecutor()` to `CALLER_RUNS`. This may cause a deadlock in the
client in the following way:

- response with `Data` that contains compact serialized data is
processed on either IO or response thread
- when the client doesn't have the schema for the `Data`, it fetches the
schema from the cluster, using `CALLER_RUNS` means that the client
invocation for `Client.FetchSchema` is executed on the IO/response
thread and it blocks. If the response to that is also to be executed on
the same thread it results in a deadlock. See the test cases for
reproducers.

Using `defaultExecutor()` instead restores the behaviour we had before
`5.3`.

Release note: Fix client deadlock under certain conditions when compact
serialization is used.

GitOrigin-RevId: 0373f423f750779a02de3fd109561177fd5cd6dd