Suppress frequent failure log messages in RaftOp [HZG-159] (#3498)
It is possible, that due to a user error during the CP Subsystem reset,
a client continues to make requests to a group that no longer exists,
resulting in failure logs on the members. If these logs are frequent
they either hide other logs, e.g. due to frequent log rotation or the
machine may run out of space.

This PR adds `FrequentLogSuppressor` that, based on the parameters,
reduces the number of identical failure logs.

- `logPeriod` - default 60 s, how often a log suppressed log should be
printed
- `threshold` - default 10, how many log failures to print before
suppressing

These parameters are exposed as internal properties.

It also fixes the propagation of the `groupId` from the member to the
client when `CPGroupDestroyedException` is thrown.

GitOrigin-RevId: a4e7bebb1635278d31a4b02742afc224e5f5f3aa