Deregister Node from TestNodeRegistry on instance shutdown (#3351)
When a HazelcastInstance is shut down (gracefully or forcefully), its reference is automatically removed from HazelcastInstanceFactory. However, in mock network, its Node is not automatically removed from TestNodeRegistry on shutdown. In the current state, TestHazelcastInstanceFactory provides `terminate(HazelcastInstance)` to notify TestNodeRegistry of the shutdown, which is inconvenient since it doesn't cover shutdowns initiated by internal means. To provide a unified experience for real and mock networks, this changeset

1. adds a lifecycle event listener to HazelcastInstance that deregisters Node on shutdown, and
2. removes `TestHazelcastInstanceFactory#terminate(HazelcastInstance)` (along with `TestNodeRegistry#removeInstance(Address)`) and replaces its usages with `TestUtil#terminateInstance(HazelcastInstance)`.

GitOrigin-RevId: 3e8fa8115c4ce217b788bf3a936173116fba3da7