Remove empty custom-lib directory from full distribution (#3545)
After removing Hazelcast 3 connector in #25051 we kept the `custom-lib`
directory in the root of our distribution, but only for the full
distribution.

This removes the directory, unifying the full and slim distribution in
this regard.

Also, removes `-Djet.custom.lib.dir=$HAZELCAST_HOME/custom-lib`
parameter given to the Hazelcast process. This means the `custom-lib`
will be looked up in current directory, not HAZELCAST_HOME directory.
This is the same behaviour we have for `cp-data` dir and persistence
`baseDir`.

GitOrigin-RevId: 02e60e1df3c83dda5c8c24a6b75c7c50e30194ba