Split brain support for VectorCollection [AI-146] [AI-151] [AI-216] (#3483)
- Adds configuration in `VectorCollectionConfig` for referencing split
brain protection name and split brain merge policies.
A `SplitBrainPolicyAwareConfig` interface is introduced
that defines the contract for a configuration class that supports
split-brain protection and merge policies. It was not applied to existing
config classes other than `CardinalityEstimatorConfig` (as a sample). If this
seems like a good idea, it can be applied to each config class that supports
split-brain related configuration. With this abstraction in place, it is also easier
to reuse other pieces of code (e.g. split-brain-related
configuration-to-XML support in `DynamicConfigXmlGenerator`).
- Adds support for split brain protection in `VectorCollectionService`.
- Adds merge operation that supports built-in generic merge policies
`DiscardMergePolicy`, `PassThroughMergePolicy` and
`PutIfAbsentMergePolicy` as well as user-supplied merge policies that
support `VectorCollectionMergeTypes`.

Fixes https://hazelcast.atlassian.net/browse/AI-146,
https://hazelcast.atlassian.net/browse/AI-151,
https://hazelcast.atlassian.net/browse/AI-216

Public API Changes:
* API: Adds split brain protection name and merge policy config in
`VectorCollectionConfig`.
* client protocol format: Adds split brain protection name and merge
policy config in `DynamicConfigVectorCollectionConfigCodec`.

---------

Co-authored-by: Krzysztof Jamróz <79092062+k-jamroz@users.noreply.github.com>
GitOrigin-RevId: 527d1e2ba3ad10dcb05130cbe8d4bf228c995520