Fix task config publishing in Kafka Connect [HZG-220] (#3564)
Two issues result in a processor not starting a task, both related to
the task config publishing.

First issue in `LateJoiningListener` is that when
`ringbuffer.tailSequence()` returns -1, the initialSequence is
initialized to `ringbuffer.tailSequence() + 1` in
`com.hazelcast.topic.impl.reliable.MessageRunner`. If a message is
published between the two calls to `tailSequence()` it will be missed.
Returning zero, instead of -1 ensures the message is not lost (there is
no 2nd call to `tailSequence()`.

The second issue is a race in `SourceConnectorWrapper`. The
`taskConfigPublisher` registers the listener and then the
`sourceConnector` and `taskRunner` are created. It is possible that the
message is received and processed before these fields are set, resulting
in a hidden NPE. Registering the listener only after these fields are
set fixes the issue.

Also adds log for any exception in `processMessage` in order to log the
NPE. Changes some log levels from `info` to `fine` level as they are
quite verbose for normal operation.

Fixes https://github.com/hazelcast/hazelcast/issues/26114

Co-authored-by: Tomasz Gawęda <tomasz.gaweda@outlook.com>
GitOrigin-RevId: a709fda4612816673636d5213c5633ad0975f2b5