Automate `release-notes.txt` generation [DI-265] (#3381)
Currently, we are manually updating the contents of `release-notes.txt`
[each release](https://github.com/hazelcast/hazelcast-mono/pull/3223) -
which is simply a versioned URL linking to the release notes - to
reflect the updated version. This is only happening on the _release_
branches, which is why in `master` it _appears_ I'm changing this
behaviour too, when I'm not.

These URLs should be generated dynamically.

Changes:
- Convert `release-notes.txt` to a template
- Dynamically compute using [`maven-resources-plugin` `filtering`
functionality](https://maven.apache.org/plugins/maven-resources-plugin/examples/filter.html)
- Extend existing Maven License test to cover other distributions
aspects (including this)
- Refactor the test to reduce duplication

Note - this functionality assumes that the URL format will remain
constant (as it has so far) and that the URL resolves - no validation is
performed to assert this behaviour. This is a requirement as the release
notes content is only populated at the end of a release, but _could_
cause issues in future if something goes awry.

Fixes: [DI-265](https://hazelcast.atlassian.net/browse/DI-265)

[DI-265]:
https://hazelcast.atlassian.net/browse/DI-265?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: 9ef103831985c0e2e60084f9157233ea160518c4