Implement Throwing functions (#3615)
Having a Runnable/Consumer as functional interface and having to be able
to throw in this is pretty annoying.
We have RunnableEx/ConsumerEx, but those are also Serializable and
SecuredFunctions.

This PR implements ThrowingRunnable/ThrowingConsumer, which allows
checked exceptions in method body (e.g. in lambda expression), but is no
Serializable nor SecuredFunction.

GitOrigin-RevId: 2a5adcea1228c4e266145c453014d30917bd6ec1