Refactor `NodeExtension` Classes (#3714)
We have configured checkstyle with an upper limit of 1,600 lines for a
class, which `EnterpriseNodeExtension` is dangerously near. Even adding
a [handful of lines means an exclusion needs to be
added](https://github.com/hazelcast/hazelcast-mono/pull/3447#discussion_r1882886911)
which is a clear sign the size needs to be reduced.

This PR refactors to achieve that, and fixes some Sonar warnings along
the way.

Changes:
- centralised `constructBuildString` duplicated implementations into
`BuildInfo`
- add generic types where possible
- use parameterized logging where possible
- access static members via base, not derived, class
- merge nested `if`s
- make methods `static` where risk free
- convert listener invocation to `forEach` lambdas
- fix some odd `catch` formatting with double spaces
GitOrigin-RevId: d9b1dd0c3a3c0dbf218bf6a8fb1f6234e7431cb5