UCN support for merge policy [CORE-186] (#3616)
https://hazelcast.atlassian.net/browse/CORE-186

Added support for merge policies UCN.

In MergeOperation, the NamespaceUtils.callWithNamespace method is used
to deserialize the merge policy. An alternative approach could be
passing the merge policy as a class name string instead of an object and
creating its instance in the @beforeRun method. This change might be
preferable, as passing a string could result in faster serialization and
deserialization. (ask reviewers opinion)

Hazelcast structures and processes with merge operation:
- [x]  Imap
- [x]  multimap
- [x]  replicated map
- [x]  icache
- [x]  iqueue
- [x]  ringbuffer
- [x]  ISet
- [x]  IList
- [x]  ScheduledExecutor
- [x]  Cardinality -  unchanged as it does not use namespace
- [x]  **Vector - to be done in separate pr**
- [x] WAN

(?) WanMapAddOrUpdateEvent remains unchanged as it is used by the OS
version, which does not include namespace support.

merge after https://github.com/hazelcast/hazelcast-mono/pull/3640

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases
- [ ] Fix documentation

---------

Co-authored-by: Jack Green <JackPGreen@Gmail.com>
Co-authored-by: Tomasz Gawęda <tomasz.gaweda@outlook.com>
GitOrigin-RevId: 3e01285bb270005b2fa8097c8b6dda3e99ea11d6