Remove unused `CPGroupIdCodec` (#3771)
`CPGroupIdCodec` was added in
https://github.com/hazelcast/hazelcast-mono/pull/591, however the
corresponding [client protocol
change](https://github.com/hazelcast/hazelcast-client-protocol/pull/504)
never generated a `CPGroupIdCodec`.

I _believe_ it was introduced in error.

Because this file has a generated hash, and we check the generated hash
as a test, [we are unable to increment the copyright year on the
codebase](https://github.com/hazelcast/hazelcast-mono/pull/3765) as
regenerating the client-protocol does not update this class, causing a
[test
failure](https://github.com/hazelcast/hazelcast-mono/runs/35038584480).

GitOrigin-RevId: 30de521f862e7eea67f35d3922ecdbefac8cd661