Correct includeValue behaviour for local IMap listeners [HZG-288] (#3790)
During a PoC with a customer, it was discovered that local listeners
were sending values to listeners even if configured to not include
values - this was likely done due to the fact that local listeners have
less overhead, but this creates confusing behaviour for customers and
does still include some degree of overhead.

This PR corrects this behaviour by checking the `isIncludeValue` of the
listener config when initializing local listeners.

This PR also includes tests to validate behaviour works as expected.

Fixes https://hazelcast.atlassian.net/browse/HZG-288

GitOrigin-RevId: 8d1119dfa043daacc866b16f02e8256096b85472