Update equals and hashCode implementation for vector collection classes [AI-278] (#3815)
Compares `hints` in `SearchOptionsImpl.equals`. Uses standard
implementation of hashCode when it does not matter - for classes that
are not supposed to be hashed. Adds standard validation of equals and
hashCode methods.

Breaking changes:
* implementation of `equals` and `hashCode` changed only for classes
which are not supposed to be used and Map keys etc., so it should not
have any impact.

Fixes: https://hazelcast.atlassian.net/browse/AI-278
GitOrigin-RevId: 33862e63bee26712ec6e094b804386b6ef6893f4