Add support for blocking backup operations [AI-192] (#3846)
This PR introduces support for blocking and optionally offloaded backup
operations.

### Blocking backup operations

Up until now the only available mechanism for long-running backup
operations was offloading, optionally using Step Engine implemented for
IMap. Blocking backup operations introduced in this PR use some of Step
Engine support code (in particular delayed sending of backup ACK) and
use `OperationParker` to implementing waiting. `OperationParker` was
extended to be aware of backup operations. Before its logic was well
suited only for primary operations.

Blocking backup operation have some unique traits:
- backup operation should not be cancelled when original member or
client disconnects, unlike for primary operations
- they can be parked both on source and destination on migration.
Primary operations can ever be parked only on migration source - they
execute on the owner and owner is never a destination of migration.
- some migrations are backup->backup (changing replica index). In some
such cases parked operations should be kept on the migration source (but
not destination) as partition data is not updated by migration on
source. This uses similar checks to the ones determining if the
partition data should be retained or deleted after migration commit and
is implemented by `BlockingBackupOperation.shouldKeepAfterMigration`
- the operations should be properly cleared during migration to avoid
executing stray operations when the partition/replica is migrated again
to the same node
- the operations should be cleared also after replica sync on
destination
- if the operation is rejected from execution (eg. by
`metWithPreconditions` or other exception) it should mark replica as
dirty to trigger anti-entropy task (note that it is not triggered
immediately). This is similar to offloaded backup operations. This is
the last resort if the backup operation cannot be executed properly and
in order. This should happen rarely and is quite costly as it triggers
replication of entire fragment partition (usually partition of single
data structure).
- backup operations must preserve order or alternatively be commutative
(rare case). This affects requirements on `shouldWait` implementation -
it is not allowed to return `true` for given operation once it returned
`false`
- because of the ordering requirements all non-commutative blocking
backup operations for given data structure partition must block on the
same condition and `WaitNotifyKey`

Additionally blocking backups can also be offloaded, but offloaded
backups need to take care for ACK sending and error handling on their
own if not using Step Engine.

With blocking backup operations `isClusterSafe`/`isMemberSafe` checks
are not fully reliable. They will return true even if there are some
parked blocking backups. Such backup operations can be lost in case of
owner crash. This is hard to observe and can happen only for invocations
that did not get their backup ACK in time.

Note: proper discarding of backup operations during migration and
promotion and ensuring that the replicas are not silently out-of-sync is
the most tricky and hard to test part of this change.

### Impact on existing workloads

Currently no data structure uses blocking backups, so there will be no
direct impact.

However some impact on migration and replication performance is possible
in cases with large number of blocking operations. After the change the
list of parked operations will have to be traversed in more cases than
before, not only after migration on destination (as originally) but also
on source and after replica sync on backup. `WaitSet` is not partitioned
based on partitionId or replicaIndex or namespace.

Part of https://hazelcast.atlassian.net/browse/AI-192

GitOrigin-RevId: 4cbf961ac0d7ddb88448c8a00961ba62bc925dd6