Do not compensate for lost backup members if fail on intermediate state is enabled (#3812)
Mechanism introduced in
https://github.com/hazelcast/hazelcast/pull/12213 (see also
https://hazelcast.atlassian.net/wiki/spaces/IMDG/pages/303598429/Slow+client+reaction+when+backup+owner+leaves+cluster)
is too aggressive for invocations with
`FAIL_ON_INDETERMINATE_OPERATION_STATE` enabled. Backup compensation
adds 1 backup ack for each invocation waiting for backup acks when a
member leaves the cluster without discriminating on which member backups
of given operation are still executing. This decreases durability of
writes (invocation may finish before all sync backups are executed), but
it does not significantly worsen current situation without
`IndeterminateOperationStateException` enabled in which some backups
still could be silently lost (an later fixed by anti-entropy
mechanisms). However, this is detrimental to
[`IndeterminateOperationStateException`
guarantees](https://docs.hazelcast.com/hazelcast/5.5/architecture/data-partitioning#throwing-an-indeterminateoperationstateexception)
as the exception may not be thrown in cases when it should be thrown:

> if an invocation completes without
`IndeterminateOperationStateException` when the configuration is
enabled, it is guaranteed that the operation has been executed
exactly-once on the primary replica and specified number of backup
replicas of the partition

The scenarios seem to be rare, but possible, eg:

1. backup operation is lost due to network problems and an unrelated
member (neither primary nor backup) leaves the cluster at the same time
2. unrelated member and backup member are lost at the same time.
3. backup is slow (>5s) which would trigger the exception but unrelated
member leaves the cluster and the exception is not thrown.

`IndeterminateOperationStateException` is among others used by Jet to
improve consistency of job snapshots.

GitOrigin-RevId: abf607dd00f203c9a9833bfd9566de29fb0f2d11