Migrate `HazelcastCommandLineTest` from `SNAPSHOT` dependencies (#3886)
`HazelcastCommandLineTest` uses a `5.0-SNAPSHOT` Hazelcast dependency,
which some analysis tools flag as being vulnerable to
[CVE-2022-36437](https://nvd.nist.gov/vuln/detail/CVE-2022-36437):

![image](https://github.com/user-attachments/assets/5372d3d6-0d37-4e65-8b0f-99c93e608811)

As it's test code this isn't an issue, but it _should_ be using a
released version (if it's deliberately an older version). And if were
incrementing the version, we might as well move to a version that's not
flagging up this warning.

[Slack
discussion](https://hazelcast.slack.com/archives/C01LH0K3VK8/p1737974870801709).

GitOrigin-RevId: 001553eac6f7fc3b025ef426942caebbb7877385