Update ArchUnit test for internal class exposure (#3885)
The ArchUnit test introduced in [this
PR](https://github.com/hazelcast/hazelcast-mono/pull/3565) to prevent
the exposure of internal classes only applies to config classes and only
checks for `.impl.` packaged classes. We should apply this to all public
API classes, and check for `.internal.` packaged classes as well.

This PR makes the updates required for this, applying the check to all
classes that are not in `.internal.`, `.impl.`, `.util.`, or `.test.`
packages and do not contain `Util`, `Test`, or `Impl` in their class
names. Various rule exceptions are also added to account for areas not
changed due to backwards compatibility.

GitOrigin-RevId: 61b08a7a294d7454459c415898dcbf70de82eb5b