Introduce Diagnostics Config [DEX-213] (1/4) (#3713)
This PR introduces the first part of diagnostic service enablement at
runtime. Currently it covers the `DiagnosticsConfig` object which can be
used by both client and member diagnostics services. The only public api
introduced in the PR is the `DiagnosticsConfig` and `DiagnosticsOutputType` classes. The rest of the
work will be handled upcoming PRs due to sake of simple reviewing.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases

---------

Co-authored-by: celenmelike <melike.celen@hazelcast.com>
GitOrigin-RevId: 312149597a321833429067c168e55b32d3959b5e