Update client-side CP stub impl for OS [HZG-316] (#3940)
Following on from the changes made in PR #3927, it was noted that those
changes only apply to older HZ clients which still implement the
original CP message tasks, but newer clients (released after the CP move
to EE), use a stub implementation of `CPSubsystem` which throws an
exception without even attempting to send a client message to the
cluster.

This PR updates the `CPSubsystemStubImpl` to be able to define whether
the stub implementor is a client or a member, and produce a different
exception message for each. This resolves the confusion about whether
the client or member is at fault for CP subsystem access failure.

Fixes https://hazelcast.atlassian.net/browse/HZG-316
Related to https://github.com/hazelcast/hazelcast-mono/pull/3927

GitOrigin-RevId: fe723565f3be86f3e6e46d2beb55349afa437bd4