More Spring tests, expose Hazelcast objects as beans [DEX-317] (#3842)
This PR:
- Adds more tests around Java configuration
- Adds test using Spring Boot
- Add exposing Hazelcast objects as Spring beans

The last one is important for consistency. When user used old XML
configuration, he could use `hz:map` tag and declare map, then
`@Autowired` this IMap as field. Since in Java configuration it's not
needed to use custom tags, a way to expose maps etc. as beans was needed

---------

Co-authored-by: Jack Green <JackPGreen@Gmail.com>
GitOrigin-RevId: e3370d36658943b715bc6db2c44ba62010aa0309