Fail-fast with HotRestart when >1 node crashes [HZG-312] (#3961)
Hot Restart currently supports the following scenarios:
- 1 node crash
- Whole cluster crash
- Rolling restart

Although a 2 node crash scenario is not supported, we can run into
unforeseen problems due to the current Hot Restart (HR) mechanics. HR
stores a snapshot of the partition table when a member crashes, so that
it can be used upon that member rejoining the cluster to speed up
recovery. The problem is that when more than 1 node crashes at the same
time, we store a snapshot for each crashed member, but the partition
table view is no longer accurate. This can lead to a situation where a
crashed node is restarted and loads data from disk into its own local
partitions, but then receives a partition table update from `master`
after startup which shifts some of its data into partition replicas
which are no longer referenced. This causes excessive memory usage
(particularly noticeable for `NATIVE` memory data) which cannot be
reclaimed, and is transferred during migration.

This PR resolves this scenario by simply keeping track of crashed nodes,
and should a node crash when 1 or more other nodes are marked as
crashed, it is marked as "unsafe", which results in the `master` node
rejecting these nodes if they attempt to load data from disk and rejoin
the cluster. They intentionally fail-fast with a `HotRestartException`,
but should be able to startup successfully on the second attempt if
`auto-remove-stale-data=true`. The supported case of 1 node crashing is
preserved.

Several test cases are included to ensure correctness under various
scenarios. It was also noted that several flaky `SlowTest` tests exist,
which were failing frequently on master. These are also resolved in this
PR, specifically within this commit:
https://github.com/hazelcast/hazelcast-mono/pull/3961/commits/9e97a635e610515d204028da37ee80188d14a063

A reproducer for the original issue (which requires killing the java
processes) can be found int [this SUP
ticket](https://hazelcast.atlassian.net/browse/SUP-769).

Fixes: https://hazelcast.atlassian.net/browse/HZG-312
GitOrigin-RevId: d6cd6826b7353d5a627d879cf8addd65e7575344