Derive sts names from local pod name instead of serviceName config [HZG-324] (#4010)
Fixes issue introduced in [this
PR](https://github.com/hazelcast/hazelcast-mono/pull/3473) where the
platform was updated to use the `serviceName` field in the k8s config to
resolve the sts it needs to monitor. There is no relation between the
service name and the sts name in general even though there is while
using the operator. This led to an issue observed in Hazelcast cloud.

The sts name can easily be derived from the local pod name as it was
before. From this sts name we can resolve the sibling sts name and
monitor both of them.

Fixes [HZG-324](https://hazelcast.atlassian.net/browse/HZG-324),
[HZC-7184](https://hazelcast.atlassian.net/browse/HZC-7184)

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Add `Add to Release Notes` label if changes should be mentioned in
release notes or `Not Release Notes content` if changes are not relevant
for release notes

[HZG-324]:
https://hazelcast.atlassian.net/browse/HZG-324?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[HZC-7184]:
https://hazelcast.atlassian.net/browse/HZC-7184?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: 2acee0d32de61b03f0328b0a1cfe1ffbbb842794