Add interceptGet() for read from backup op [HZG-331] (#3998)
The MapInterceptor does not currently call the `interceptGet()` method
when reading from backup data is enabled on an `IMap`, however
`afterGet` is called.

This PR makes the simple change to have the interceptor called for read
from backup `get()` calls, and also adds a test to verify this
functionality.

Fixes https://hazelcast.atlassian.net/browse/HZG-331

GitOrigin-RevId: c72b9e302e6b19d51951322382069731065eb002