Fix `MCEnterpriseMessageTasksTest` line separators assertions on Windows [HZG-238] (#4083)
The test added in https://github.com/hazelcast/hazelcast-mono/pull/3241
introduced a test failure when a file with hardcoded line separators was
compared against a dynamic method.

Note - there are _many_ possible solutions to this problem, I tried to
use one that's _already_ in use elsewhere, even if potentially not the
best.

Fixes: https://github.com/hazelcast/hazelcast-mono/issues/3383,
[HZG-238](https://hazelcast.atlassian.net/browse/HZG-238)

[HZG-238]:
https://hazelcast.atlassian.net/browse/HZG-238?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: fa2d2bdfeadeab2f32db200e1ba5f85e244e6015