Support env variable property declaration for SSL [HZG-200] (#4085)
The specific issue with using environment variables for configuring SSL
properties is that we turn all property keys to lowercase when
processing (for various reasons, mainly because all our config options
are lowercase, so it makes sense for 99% of other cases). This breaks
the functionality of SSL and means you can't use SSL if properties are
declared through env variables (such as used through Docker, as was the
case in HZG-200).

This PR fixes this by checking for a lowercase variant of the key when
reading properties in `SSLEngineFactorySupport`. This resolves the issue
for SSL properties specifically, which is the only concern currently
raised for this feature.

Fixes https://hazelcast.atlassian.net/browse/HZG-200

GitOrigin-RevId: 7196556ed51a2796b9e72813505731c601190ac1