Ensure `EnterpriseTimedMemberStateFactoryTest` System property changes are isolated [DI-441] (#4097)
The test clears system properties, but does not use any fixtures that
isolates system property mutation.

This caused `ScheduledExecutorServiceBasicTest` to fail as if executed
without a license because of the `System.clearProperty` - but only on
Jenkins in AWS.

After extensive debugging, was unable to properly identify the root
cause, but believe the issue is due to an incompatibility between how
JUnit5 tests are executed, and the (known leaky) nature of Hazelcast
JUnit Enterprise runners.

The simplest fix is to covert this to a JUnit4 test using the same
runners as used in other tests, and to extend the existing tests to flag
_clearing_ a system property (not just setting) as requiring
encapsulation.

Fixes: [DI-441](https://hazelcast.atlassian.net/browse/DI-441),
https://github.com/hazelcast/hazelcast-mono/pull/4091

[DI-441]:
https://hazelcast.atlassian.net/browse/DI-441?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

GitOrigin-RevId: dc1dd52fcf2cf0ea0e823c5d6c2caa28d0aeb1c6