Correct ByteArrayObjectDataOutput capacity for large inputs [HZG-346] (#4099)
There is a demonstrable large performance hit (100x) when an input
length of more than `Integer.MAX_VALUE / 2` is provided. This is due to
resizing being done in increments of `len` passed to
`ensureAvailable(int len)` when doubling the current buffer length with
a bit shift would overflow and become a negative value.

This means that the buffer is potentially continuously extended by small
increments, instead of a single large increment (which is the intention
with a doubling strategy).

This PR resolves the issue by catching the overflow case and setting the
buffer's length to a `MAX_ARRAY_SIZE` value, which is approximately the
largest supported array size possible.

This is the simplest solution to the problem encountered, and fits with
the current doubling strategy within the limits of the JVM. We could
overhaul our buffering mechanism entirely and use a segmented approach,
trading some additional overheads for a more robust solution that
supports larger inputs - this is definitely a more complex solution, and
is likely overkill for Hazelcast's use-case. I discussed these options
with @gbarnett-hz and he agreed that the simple approach is sufficient
for now.

This PR also includes a test case that increases the existing
`ByteArrayObjectDataOutput` test coverage.

Fixes https://hazelcast.atlassian.net/browse/HZG-346
Closes https://github.com/hazelcast/hazelcast/issues/26422

GitOrigin-RevId: aaf54ed54a7f6c9238858e854938bdfd1d66a01a