java.lang.SecurityException: setContextClassLoader [HZG-357] (#4195)
jira - https://hazelcast.atlassian.net/browse/HZG-357

When running the application inside a Docker container with param
-Djava.security.manager=allow, the Java common ForkJoinPool creates
instances of InnocuousForkJoinWorkerThread instead of the standard
ForkJoinWorkerThread. As a result, the Thread.setContextClassLoader
throws the SecurityException.
This behavior differs from running the same application outside of
Docker, where normal ForkJoinWorkerThread instances are created, and no
exception is thrown.
Reproduced on 21.0.6.7-1, but does not reproduce on 21.0.5.11-1

Created issue on corretto, but looks like this is a problem (or feature)
of jdk.

To avoid relying on unexpected JDK behavior, this PR replaces the common
pool with a custom one.

UPD:
I was hoping all the issues could be resolved in one go, but it looks
like they’ll need to be addressed one by one.
All instances where the common pool is used along with
setContextClassLoader are affected.
However, based on the test results, it seems that only one such place is
currently impacted.

The bug was introduced in OpenJDK version 21.0.6.7.1.
Issue in openJDK Jira is created.

Checklist:
- [x] Labels (`Team:`, `Type:`, `Source:`, `Module:`) and Milestone set
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [x] Send backports/forwardports if fix needs to be applied to
past/future releases
GitOrigin-RevId: d8c390a0b903010faadfd9182ce5f666d4c90e77