Use consistent Docker image versions in tests (#4278)
Some of the tests use Docker images (e.g. for a MySQL instance).

_Some_ tests make use of a centralised `TestedVersions` class to use a
version of the Docker image that can be overridden with a system
property for testing purposes, but others have their own hardcoded
version that may or may not be overridable.

Updated to centralise all duplicated Docker image references into
`TestedVersions`.

In the case of conflicts, picked the latest version we already used
(which was the one in `TestedVersions` anyway):
- `debezium/example-mysql`
   - `2.3.0.Final`
   - `2.6.2.Final`
   - `2.7.1.Final`
- `debezium/example-postgres`
   - `2.3.0.Final`
   - `2.6.2.Final`
   - `2.7.1.Final`
- `mongo`
   - `6.0.3`
   - `7.0.5`
- `mysql`
   - `8.0.32`
   - `8.2`

GitOrigin-RevId: 7ff9cc6e7498757defc62db82eb5e9953e25762e