Include expired entries when depopulating global indexes (#4294)
Otherwise, the HD global index will have dangling pointers.

**Issue:**
Migration finalization clears the data on the migration source node, but
if an entry is expired, the cleanup was skipping the removal of the
expired data from the global index. As a result, query threads that hit
this part of the index encountered dangling pointers, which led to JVM
crashes.

**Fix:**
Do not skip expired entries when depopulating global indexes.

GitOrigin-RevId: 471b5f01d3d05e1ecf6ca1a23681c11da6b4c96e