Avoid `MethodHandles.lookup` `VarHandle` boilerplate (#4268)
When obtaining a `VarHandle` we have copy/pasted boilerplate to `catch`
exceptions etc which makes the code more verbose.

We can extract this to a utility class to make the code more concise and
reduce the risk of accidentally copy/pasting the wrong class reference
by inferring dynamically.

This also means that
https://github.com/hazelcast/hazelcast-mono/pull/4267 can be done as a
drop-in replacement.

GitOrigin-RevId: addd84ab4b88d790a4e11f9e7d464bfc77248e39