Fix QueryCacheEventData not receiving Versioned input [HZG-385] (#4354)
A customer reported to us that they were seeing some lost Continuous
Query Cache (CQC) events, with accompanying NullPointerExceptions
(NPEs). The NPE stack traces showed that while reading CQC event data,
the `mapName` field was `null` in some cases.

Introducing some extra assertions and running our nightly test builds
revealed that `mapName` was indeed being found to be null in
`DefaultQueryCacheEventData` objects when read from `BatchEventData`.

After some investigation, I identified the cause: the `ObjectDataInput`
passed to `DefaultQueryCacheEventData#readData` does not have its
version defined, so it is using version `0.0`, which means the
`RU_COMPAT_5_3` code branch in `DefaultQueryCacheEventData` was not
being reached - thus `mapName` was not read, and was left as `null`.

This happens because `BatchEventData` reads `QueryCacheEventData` within
its own `readData` method, and since `BatchEventData` does not implement
`Versioned`, this input stream did not have a version set.

This PR resolves the issue by marking `BatchEventData` as `Versioned`,
so that the event data it loads will also receive a versioned input
stream. It has been noted that `DefaultQueryCacheEventData` is also
annotated with `BinaryInterface`, and so should not have implemented
`Versioned` in the first place - but after walking through the code
paths, we believe this may be an overly cautious annotation, as client
compatibility is handled through the protocol encoding. This will be
further investigated for a `6.0.0` release, to either remove these
annotations, or correct this `Versioned` behaviour.

Test coverage for this is now handled in existing tests (such as
`QueryCacheSimpleStressTest`) with the addition of an assertion in
`BatchEventData#readData`.

Fixes https://hazelcast.atlassian.net/browse/HZG-385

GitOrigin-RevId: a1a9d45ee803264cc779c4d8107c1d4045588cb7