Immutable put interceptor inputs [HZG-381] (#4317)
jira - https://hazelcast.atlassian.net/browse/HZG-381

Fixed:

**1. Initial bug reason:**
Interceptors were able to modify the input value directly.
When the value was of type Object (non-binary), this could
unintentionally impact downstream logic after executing the interceptor.

**Solution:**
Added additional Javadoc to clarify the mutability of input values in
interceptors.
Included a comment explicitly stating that modifying the input object is
not supported when in-memory-format=OBJECT and an index is present, as
it can lead to unexpected behavior.

**2. Added tests for memory type = OBJECT.**

**3. AfterPut Behavior:** afterPut behaved differently for regular put
and for EntryProcessor.
Fix: afterPut now consistently receives the result of the
putInterceptor.

Other issues(?) (not fixed)
https://hazelcast.atlassian.net/browse/HZG-398

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases
- [x] Fix documentation - https://github.com/hazelcast/hz-docs/pull/1704

---------

Co-authored-by: James Holgate <130981049+JamesHazelcast@users.noreply.github.com>
GitOrigin-RevId: 6a7e8228a042fda9a1dcb0b2772c1e8c252fc974