master node change exception [HZG-480] (#4674)
MasterNodeChangedException extension of MemberLeftException was not a
good idea.

See these comments for context:

https://github.com/hazelcast/hazelcast-mono/pull/4636#discussion_r2193001575

https://github.com/hazelcast/hazelcast-mono/pull/4636#discussion_r2195436755

MemberLeftException has custom serialization logic that depends on a
Member object—specifically, the member that has left the cluster.
Serialization of MasterNodeChangedException fails due to the absence of
a Member object.

1) It’s not possible to override this serialization properly in
MasterNodeChangedException (java serialization restriction).
2) In the case of MasterNodeChangedException, we don’t actually know
which member has left. This exception is thrown either when a new node
starts at the same address as the previous master, or when there is no
master node yet — so there’s no specific Member to attach to the
exception.

Checklist:

- [x] Labels (`Type:`, `Source:`, `Module:`) and Milestone set
- [x] Architecture Design Record (ADR) reviewed and signed-off if this
PR represents a significant architectural change
- [x] Request reviewers if possible
- [x] New public APIs have `@Nonnull/@Nullable` annotations
- [x] New public APIs have `@since` tags in Javadoc
- [ ] Send backports/forwardports if fix needs to be applied to
past/future releases

GitOrigin-RevId: b110ef21072878fa90f226ffd93144a9d25d022e